Use uint16 instead of uint8 for FileOptions

+++ internal/api/model/api.go
@@ -382,7 +382,7 @@ type UpdateAllocationRequest struct {
 	AddBlobberId         string `json:"add_blobber_id"`
 	RemoveBlobberId      string `json:"remove_blobber_id"`
 	ThirdPartyExtendable bool   `json:"third_party_extendable"`
-	FileOptions          uint8  `json:"file_options"`
+	FileOptions          uint16  `json:"file_options"`
 }
 
 type SCRestGetBlobberRequest struct {
